requirements.txt
streamlit
yt-dlp
static-ffmpeg
Pillow